requirements.txt
pandas==1.3.1
streamlit==1.12.2
plotly.express==0.4.1 
re==2.2.1